+++ modules/Accounting/functions.inc.php
@@ -438,15 +438,6 @@ function _saveSalariesFile( $id )
 		return '';
 	}
 
-	$file_name_no_ext = no_accents( mb_substr(
-		$_FILES[ $input ]['name'],
-		0,
-		mb_strrpos( $_FILES[ $input ]['name'], '.' )
-	) );
-
-	// @since 11.0 Add microseconds to filename format to make it harder to predict.
-	$file_name_no_ext .= '_' . date( 'Y-m-d_His' ) . '.' . substr( (string) microtime(), 2, 6 );
-
 	$file_attached = FileUpload(
 		$input,
 		$FileUploadsPath . UserSyear() . '/staff_' . UserStaffID() . '/',
@@ -454,7 +445,7 @@ function _saveSalariesFile( $id )
 		0,
 		$error,
 		'',
-		$file_name_no_ext
+		FileNameTimestamp( $_FILES[ $input ]['name'] )
 	);
 
 	// Fix SQL error when quote in uploaded file name.
@@ -497,15 +488,6 @@ function _saveIncomesFile( $id )
 		return '';
 	}
 
-	$file_name_no_ext = no_accents( mb_substr(
-		$_FILES[ $input ]['name'],
-		0,
-		mb_strrpos( $_FILES[ $input ]['name'], '.' )
-	) );
-
-	// @since 11.0 Add microseconds to filename format to make it harder to predict.
-	$file_name_no_ext .= '_' . date( 'Y-m-d_His' ) . '.' . substr( (string) microtime(), 2, 6 );
-
 	$file_attached = FileUpload(
 		$input,
 		$FileUploadsPath . UserSyear() . '/staff_' . User( 'STAFF_ID' ) . '/',
@@ -513,7 +495,7 @@ function _saveIncomesFile( $id )
 		0,
 		$error,
 		'',
-		$file_name_no_ext
+		FileNameTimestamp( $_FILES[ $input ]['name'] )
 	);
 
 	// Fix SQL error when quote in uploaded file name.